@@ -1,18 +1,17 @@
 name: flutter_audio_query
 description: Flutter plugin to query data about artists, albums, songs, genres and playlists
   from device storage.
-version: 0.3.5+6
+version: 1.0.0+1
 #author: Marcos Antonio Boaventura <scavenger.gnu@gmail.com>
 homepage: https://github.com/sc4v3ng3r/flutter_audio_query
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
-  flutter: ">=1.10.0 <2.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
-  flutter_plugin_android_lifecycle: ^1.0.6
+  flutter_plugin_android_lifecycle: ^2.0.0
 
 
 flutter:
